Creatd files to separate logic and dom
